refactor: 동의 내용 - 상수 폴더로 이동
